cleaning component & rewriting some parts
